developing banner.ejs
Banner is added as a partial and could be called anywhere it is needed.
Calling method: Put your text in "msgJson" and add this tag on your ejs file,  <%- include('../partials/banner', { "msgJson": "", "maxCharLengthJson" :218 , "freezTimeJson" : 110 , "colorCodeJson" :"#fcba03"} ) -%>

This feature has 4 parts as input that should be sent as JSON
First part is the Text, that you want to show it.
second part is the maximum character that present the length of your banner
-Third part is the freezing speed, and you can setup speed of moving text in banner
-Forth part is the background color of banner, that could be defined as any standard method in CSS/HTML